Moved inputs as state & signin as prop. Categorised it under components folder
