--- Inputs.js
@@ -1,69 +1,0 @@
-import React, { Component } from 'react'
-import { View, Text, TouchableOpacity, TextInput, StyleSheet } from 'react-native'
-
-class Inputs extends Component {
-   state = {
-      username: '',
-      password: ''
-   }
-   handleUsername = (text) => {
-      this.setState({ username: text })
-   }
-   handlePassword = (text) => {
-      this.setState({ password: text })
-   }
-   login = (username, pass) => {
-      alert('username: ' + username + ' password: ' + pass)
-   }
-   render(){
-      return (
-         <View style = {styles.container}>
-            <TextInput style = {styles.input}
-               underlineColorAndroid = "transparent"
-               placeholder = "Username"
-               placeholderTextColor = "#fff"
-               autoCapitalize = "none"
-               onChangeText = {this.handleUsername}/>
-            
-            <TextInput style = {styles.input}
-               underlineColorAndroid = "transparent"
-               placeholder = "Password"
-               placeholderTextColor = "#fff"
-               autoCapitalize = "none"
-               onChangeText = {this.handlePassword}/>
-               
-            <TouchableOpacity
-               style = {styles.submitButton}
-               onPress = {
-                  () => this.login(this.state.email, this.state.password)
-               }>
-               <Text style = {styles.submitButtonText}> Submit </Text>
-            </TouchableOpacity>
-         </View>
-      )
-   }
-}
-export default Inputs
-
-const styles = StyleSheet.create({
-   container: {
-      paddingTop: 23
-   },
-   input: {
-      margin: 15,
-      height: 40,
-      fontFamily: 'Helvetica Neue',
-      fontSize: 19,
-      borderBottomWidth: 1,
-      borderColor: '#fff'
-   },
-   submitButton: {
-      backgroundColor: '#7a42f4',
-      padding: 10,
-      margin: 15,
-      height: 40,
-   },
-   submitButtonText:{
-      color: 'white'
-   }
-})
