+++ src/components/routes/Routes.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Router, Scene } from 'react-native-router-flux'
 import StartUp from '../../../Start-up.js'
-import SignIn from '../../../Sign-in.js'
+import SignIn from '../sign-in/Sign-in.js'
 import Menu from '../../../Menu.js'
 
 const Routes = () => (
