+++ src/components/sign-in/Sign-in.js
@@ -1,0 +1,10 @@
+import React, {Component} from 'react'
+import Inputs from './Sign-in-inputs.js'
+// import PresentationalComponent from './PresentationalComponent'
+
+const SignIn = () => {
+  return (
+     <Inputs />
+  )
+}
+export default SignIn
